Move payment-method-options to inside tokenize-payment-method
